pyproject.toml
[tool.poetry]
name = "project"
version = "0.1.0"
description = ""
authors = ["Liz Dahlstrom <1231540+tanglisha@users.noreply.github.com>"]
readme = "README.md"

packages = [
    {include = "src"},
    {include = "config"}
]

[tool.poetry.dependencies]
python = "^3.12"
numpy = "^2.3.2"
pandas = "^2.3.1"
google-cloud-storage = "^3.2.0"
scikit-learn = "1.6.1"
pyyaml = "^6.0.2"
pydantic = "^2.11.7"
imbalanced-learn = "^0.13.0"
lightgbm = "^4.6.0"
pydantic-yaml = "^1.5.1"
mlflow = "^3.2.0"


[tool.poetry.group.dev.dependencies]
pytest = "^8.4.1"
ipykernel = "^6.30.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
